Move Exports functional area to its own package
