+++ src/le_beta_vis/frontend/viewmodels/HistoricalExportViewModel.py
@@ -22,12 +22,12 @@ from pathlib import Path
 from typing import Callable, List, Optional
 
 from le_beta_vis.common.AppInfo import APP_VERSION
-from le_beta_vis.common.ClusterExportService import ClusterMetadataLabels
 from le_beta_vis.common.Colormap import Colormap
 from le_beta_vis.common.ConfigurationService import ConfigurationService
 from le_beta_vis.common.EPSDataClasses import ClusterQueryFilter
-from le_beta_vis.common.ExportStorageService import CancelToken, ExportProvenance
-from le_beta_vis.common.HistoricalExportService import (
+from le_beta_vis.export.ClusterExportService import ClusterMetadataLabels
+from le_beta_vis.export.ExportStorageService import CancelToken, ExportProvenance
+from le_beta_vis.export.HistoricalExportService import (
     ExportRequest,
     HistoricalExportService,
 )
