+++ tests/test_H5ExportStorageService.py
@@ -11,8 +11,8 @@ import pytest
 
 from le_beta_vis.common.BoundingBox import BoundingBox
 from le_beta_vis.common.Cluster import Cluster
-from le_beta_vis.common.ExportStorageService import CancelToken, ExportProvenance
-from le_beta_vis.common.H5ExportStorageService import (
+from le_beta_vis.export.ExportStorageService import CancelToken, ExportProvenance
+from le_beta_vis.export.H5ExportStorageService import (
     CLUSTER_COLUMNS,
     H5ExportStorageService,
 )
