+++ tests/test_HistoricalExportService.py
@@ -12,18 +12,18 @@ import numpy as np
 
 from le_beta_vis.common.BoundingBox import BoundingBox
 from le_beta_vis.common.Cluster import Cluster
-from le_beta_vis.common.ClusterExportService import (
+from le_beta_vis.common.Colormap import Colormap
+from le_beta_vis.common.EventRepository import EventRepository
+from le_beta_vis.export.ClusterExportService import (
     ClusterExportService,
     ClusterMetadataLabels,
 )
-from le_beta_vis.common.Colormap import Colormap
-from le_beta_vis.common.EventRepository import EventRepository
-from le_beta_vis.common.ExportStorageService import (
+from le_beta_vis.export.ExportStorageService import (
     CancelToken,
     ExportProvenance,
     ExportStorageService,
 )
-from le_beta_vis.common.HistoricalExportService import (
+from le_beta_vis.export.HistoricalExportService import (
     ExportRequest,
     HistoricalExportService,
 )
