+++ tests/test_HistoricalExportViewModel.py
@@ -12,19 +12,19 @@ import pytest
 
 from le_beta_vis.common.BoundingBox import BoundingBox
 from le_beta_vis.common.Cluster import Cluster
-from le_beta_vis.common.ClusterExportService import (
-    ClusterExportService,
-    ClusterMetadataLabels,
-)
 from le_beta_vis.common.Colormap import Colormap
 from le_beta_vis.common.EPSDataClasses import ClusterQueryFilter
 from le_beta_vis.common.EventRepository import EventRepository
-from le_beta_vis.common.ExportStorageService import (
+from le_beta_vis.export.ClusterExportService import (
+    ClusterExportService,
+    ClusterMetadataLabels,
+)
+from le_beta_vis.export.ExportStorageService import (
     CancelToken,
     ExportProvenance,
     ExportStorageService,
 )
-from le_beta_vis.common.HistoricalExportService import HistoricalExportService
+from le_beta_vis.export.HistoricalExportService import HistoricalExportService
 from le_beta_vis.common.ThumbnailLoaderService import ThumbnailLoaderService
 from le_beta_vis.frontend.viewmodels.HistoricalExportViewModel import (
     HistoricalExportViewModel,
